Updated tests for new code and fixed all errors and failing tests
